+++ src/cc/Hypertable/Lib/BlockCompressionCodecLzo.cc
@@ -118,13 +118,13 @@ BlockCompressionCodecLzo::inflate(const DynamicBuffer &input,
 
   if (header.get_data_zlength() != remaining) {
     HT_ERRORF("Block decompression error, header zlength = %d, actual = %d",
-              header.get_data_zlength(), remaining);
+              (int)header.get_data_zlength(), (int)remaining);
     HT_THROW(Error::BLOCK_COMPRESSOR_BAD_HEADER, "");
   }
 
   uint32_t checksum = fletcher32(msg_ptr, remaining);
   if (checksum != header.get_data_checksum()) {
-    HT_ERRORF("Compressed block checksum mismatch header=%d, computed=%d",
+    HT_ERRORF("Compressed block checksum mismatch header=%u, computed=%u",
               header.get_data_checksum(), checksum);
     HT_THROW(Error::BLOCK_COMPRESSOR_CHECKSUM_MISMATCH, "");
   }
